requirements.txt
ibm-watson
requests
PyMuPDF
python-dotenv